refactor: exposed library through an interface file
