+++ example/integration_test/plugin_integration_test.dart
@@ -6,7 +6,7 @@
 // For more information about Flutter integration tests, please see
 // https://docs.flutter.dev/cookbook/testing/integration/introduction
 
-import 'package:flutter_security_toolkit/threat_detection_center.dart';
+import 'package:flutter_security_toolkit/src/threat_detection_center.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
