+++ lib/flutter_security_toolkit.dart
@@ -1,0 +1,3 @@
+library flutter_security_toolkit;
+
+export 'src/threat_detection_center.dart';
