+++ test/flutter_security_toolkit_method_channel_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/services.dart';
-import 'package:flutter_security_toolkit/flutter_security_toolkit_method_channel.dart';
+import 'package:flutter_security_toolkit/src/flutter_security_toolkit_method_channel.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
