+++ test/threat_detection_center_test.dart
@@ -1,6 +1,6 @@
-import 'package:flutter_security_toolkit/flutter_security_toolkit_method_channel.dart';
-import 'package:flutter_security_toolkit/flutter_security_toolkit_platform_interface.dart';
-import 'package:flutter_security_toolkit/threat_detection_center.dart';
+import 'package:flutter_security_toolkit/src/flutter_security_toolkit_method_channel.dart';
+import 'package:flutter_security_toolkit/src/flutter_security_toolkit_platform_interface.dart';
+import 'package:flutter_security_toolkit/src/threat_detection_center.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
